+++ Changelog.md
@@ -66,6 +66,8 @@ example, concat streams concurrently using this.
     * `reverse'` a several times faster reverse using Storable instance
     * `intersperseByTime` intersperse a mondadic action in a stream
       periodically
+    * `intersperse` intersperse an element in between consecutive elements in
+      stream
 
 ### Deprecations